Pass a command line flag to downstream tests (#435)
The command line flag `--downstream_integration_test` that is passed to the downstream test will allow us to selectively skip tests that might not be necessary. In particular, the Aqua test for stale dependencies is unnecessary in downstream tests if we want to test package extensions.

We also reinstate the test against `InfiniteLinearAlgebra` in this PR, which was commented out.