Preserve block information in more slicing operations (#459)
This introduces a generalization of `BlockSlice` for views involving
vectors of `Block` and `BlockIndexRange`, building off of #445.

I call it `BlockedSlice` but I'm open to suggestions. At first I tried
to just generalize `BlockSlice` to be an `AbstractVector` subtype but
some operations depend on it being an `AbstractUnitRange` so I think it
makes sense to have both.

The motivation is that in
https://github.com/ITensor/BlockSparseArrays.jl it would be helpful to
preserve information about the original blocks being sliced when taking
views involving vectors of `Block` and `BlockIndexRange`. #445 drops the
blocks input in the slice operation, so it is hard to tell if a certain
slice operation was really originally blockwise.

@dlfivefifty curious to hear your thoughts on this since you wrote #445.