Introduce `eachblockaxes`, more general `blocklengths` (#476)
Introduces and exports `eachblockaxes`, which is analogous to the
redefinition of `blocksizes` introduced in #399. It outputs a (lazy)
array with a size `blocksize(A)` that when you index into it it returns
the axes of the corresponding block of `A`. It also defines a private
function `BlockArrays.eachblockaxes1` analogous to `Base.axes1` for
getting the axes of the blocks in the first dimension.

The motivation for this is related to the discussion in #446, this helps
with use cases where the blocks of a block array have non-trivial axes,
for example they themselves might be blocked, have a Kronecker
structure, have extra information like symmetry sector information
(which is an application of [graded vector
spaces](https://en.wikipedia.org/wiki/Graded_vector_space)), have labels
that need to be preserved, etc.

It also introduces a generalization of
[`blocklengths`](https://juliaarrays.github.io/BlockArrays.jl/v1.6/lib/public/#BlockArrays.blocklengths)
beyond `AbstractUnitRange{<:Integer}` to any `AbstractArray` by
outputting a (lazy) array that when you index into it returns the length
of the corresponding block.

This PR also changes the implementation of `blocksizes`. Before, it was
implemented as a lazy version of `size.(blocks(A))`. Now, it is based on
a new `ProductArray` type, where the `blocklengths` of the axes in each
dimension are stored and accessed. Then, `ProductArray` can be shared
with and used for the implementation of `eachblockaxes`. I think that
design is easier to reason about and in the case of `eachblockaxes` it
is easier to make the element type concrete.

Related to #369 and #446.