Rename `BlockedSlice` from #459 to `NoncontiguousBlockSlice` (#479)
This PR is split off from #462, it renames `BlockedSlice` introduced in
#459 to `NoncontiguousBlockSlice` and also adds more functionality and
tests so that it is as featureful as `BlockSlice`. The reason for the
name change is that in #462, this type will be used in slicing
operations such as `V[[Block(1), Block(3)]]` and `V[Block(1)[[1, 3]]]`.
In my opinion, `BlockedSlice` doesn't make a lot of sense as a name for
the latter case. The more general concept is that it is a slice object
constructed when there is a non-contiguous blockwise slice of some form
(either non-contiguous blocks or non-contiguous within a block).

I'm open to other name suggestions, `NoncontiguousBlockSlice` is the
best I could come up with and there isn't really an analogous case for
this in Base from what I've seen. Note the main reason why we can't just
use `BlockSlice` for those cases is that `BlockSlice` is an
`AbstractUnitRange` subtype and `NoncontiguousBlockSlice` covers cases
where the slices aren't ranges, so maybe `BlockSliceVector` could work
as a name to emphasize that point.

---------

Co-authored-by: Sheehan Olver <solver@mac.com>