Fix axes and size of BlockIndices (#490)
The current definition of `size(iter::BlockIndices)` was assuming the
indices were ranges (resulting from #483 where `BlockIndexRange` was
generalized to `BlockIndices` but the definition if `size` wasn't
updated accordingly), so for (non-contiguous) non-ranges it was giving
the wrong result, this PR fixes that. Additionally, previously only
`size` was defined which meant that BlockIndices where the indices had
non-trivial axes (like blocked axes) lost information, which is now
fixed and tested.