Fix ambiguity error in `similar` with 3 or more blocked axes (#494)
I found that `similar` calls such as:
```julia
using BlockArrays
r = blockedrange([2, 2])
similar(Array{Float64}, (r, r, r))
```
(with inputs of 3 or more blocked axes) lead to ambiguity errors between
the `StridedArray` and `AbstractArray` versions which output
`BlockedArray` and `BlockArray` respectively, this PR fixes that
ambiguity by defining a more specialized `similar` method only for
`AbstractBlockedUnitRange` inputs. The existing definitions accept
mixtures of blocked, non-blocked, and integer dimensions which is more
prone to ambiguities.

Related to #196 and #489.

I also noticed these definitions are not so friendly for GPU array types
but that is a separate matter...